requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-debug-toolbar==3.2.1
django-environ==0.4.5
djangorestframework==3.12.4
flake8==3.9.2
idna==2.10
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
pyshorteners==1.0.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
